requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
redis==5.0.1
psycopg2-binary==2.9.7
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.4
pyyaml==6.0.1
prometheus-client==0.17.1
structlog==23.1.0 